Document fmt::Arguments internal representation.
